Move collection selector component into left-hand side of reflex container
